DAOS-3114 vos: Consolidate timestamp cache APIs
Simplify the timestamp cache some by consolidating APIs.  There is now one
API for checking for conflicts and one API for updating the read timestamps.
This is preparation for adding more read timestamp updates for iteration
and query APIs and for adding epoch uncertainty checks.  More simplification
is probably needed but this is a step in that direction.

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>